TUM-12 - rename Data directory to Storage in Infrastructure
